Renamed everything to goblaze and deleted the appsettings file I forgot I had
