+++ README.md
@@ -1,2 +1,3 @@
-# backblaze-manager
+# Goblaze
+
 A Go application for encrypting and uploading files to a Backblaze B2 cloud storage account
