--- appconfig.json
@@ -1,3 +1,0 @@
-{
-  "applicationKey": "0000ac10cd67b60d1e4333a8cfff77aee02297a008"
-}
