--- e2etests/e2etests_test.go
@@ -1,10 +1,0 @@
-package e2etests
-
-import (
-	"github.com/jack-ohara/backblazemanager/backblazemanager"
-	"testing"
-)
-
-func TestTryThis(t *testing.T) {
-	backblazemanager.TryThis()
-}
