+++ goblaze/fileencryption/decryption/decryption.go
@@ -5,7 +5,7 @@ import (
 	"crypto/cipher"
 	"log"
 
-	"github.com/jack-ohara/backblazemanager/backblazemanager/fileencryption/passwordhasher"
+	"github.com/jack-ohara/goblaze/goblaze/fileencryption/passwordhasher"
 )
 
 func DecryptData(data []byte, passphrase string) []byte {
